moved interval logic from UI layer to business layer (where it belongs). moved traffic light into usercontrol for reusability
