+++ UI/MainWindow.xaml
@@ -3,20 +3,20 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:UI"
-        xmlns:VM="clr-namespace:UI"        
+        xmlns:local="clr-namespace:UI"         
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800">
     <Window.DataContext>
-        <VM:MainWindowViewModel/>
+        <local:MainWindowViewModel/>
     </Window.DataContext>
     <Grid>
-        <Border BorderBrush="Black" BorderThickness="1" Height="300" Width="110">
-            <StackPanel>
-                <Ellipse Height="100" Width="100" Stroke="Black" StrokeThickness="2" Fill="{Binding Top, Mode=TwoWay}"/>
-                <Ellipse Height="100" Width="100" Stroke="Black" StrokeThickness="2" Fill="{Binding Middle, Mode=TwoWay}"/>
-                <Ellipse Height="100" Width="100" Stroke="Black" StrokeThickness="2" Fill="{Binding Lower, Mode=TwoWay, PresentationTraceSources.TraceLevel=High}"/>
-            </StackPanel>
-        </Border>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition/>
+            <ColumnDefinition/>
+            <ColumnDefinition/>
+            <ColumnDefinition/>
+        </Grid.ColumnDefinitions>
+        <local:TrafficLightUserControl Grid.Column="0" x:Name="trafficLight1"/>
+        <local:TrafficLightUserControl Grid.Column="1" x:Name="trafficLight2"/>
     </Grid>
 </Window>
